ANSRM-130  Fix the validation on monitors to be able to accept decimals (#674)
* updated procedureStepMonitor Schema

* update WorkOrderTaskMonitor schema

* update api.client.generated.ts file

* fixed targetvalue type

* updated procedureStepMonitor models

* update WorkOrderTaskMonitor Models

* update api.client.generated.ts file

* update procedurestepmonitor edit form

* update workordertaskmonitor wrapper

* portal updated

* fix

* updating only to cause a re-build of the jenkins jobs

+++ MSR.Domain/Commands/CreateProcedureStepMonitor.cs
@@ -10,9 +10,9 @@ namespace MSR.Domain.Commands
         public string Description { get; set; }
         public int? MonitorListId { get; set; }
         public string ShouldBe { get; set; }
-        public float? HighTarget { get; set; }
-        public float? LowTarget { get; set; }
-        public float? Target { get; set; }
+        public decimal? HighTarget { get; set; }
+        public decimal? LowTarget { get; set; }
+        public decimal? Target { get; set; }
         public string FailAction { get; set; }
         public string SensorName { get; set; }
         public bool? SendNCREmail { get; set; }
